requirements.txt
Django==1.6.5
Pillow==2.4.0
argparse==1.2.1
beautifulsoup4==4.3.2
django-braces==1.4.0
django-ckeditor-updated==4.4.0
django-debug-toolbar==1.2.1
django-durationfield==0.5.1
django-xmlrpc==0.1.5
psycopg2==2.5.3
pyparsing==2.0.2
pytz==2014.4
six==1.7.2
sqlparse==0.1.11
wsgiref==0.1.2
